Enhancement: Extract ObjectHasAttributeTest
