+++ tests/Constraint/ObjectHasAttributeTest.php
@@ -1,0 +1,68 @@
+<?php
+/*
+ * This file is part of sebastian/phpunit-framework-constraint.
+ *
+ * (c) Sebastian Bergmann <sebastian@phpunit.de>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace PHPUnit\Framework\Constraint;
+
+use PHPUnit\Framework\ExpectationFailedException;
+use PHPUnit\Framework\TestCase;
+use PHPUnit\Framework\TestFailure;
+
+class ObjectHasAttributeTest extends TestCase
+{
+    public function testConstraintObjectHasAttribute()
+    {
+        $constraint = new ObjectHasAttribute('privateAttribute');
+
+        $this->assertTrue($constraint->evaluate(new \ClassWithNonPublicAttributes, '', true));
+        $this->assertFalse($constraint->evaluate(new \stdClass, '', true));
+        $this->assertEquals('has attribute "privateAttribute"', $constraint->toString());
+        $this->assertCount(1, $constraint);
+
+        try {
+            $constraint->evaluate(new \stdClass);
+        } catch (ExpectationFailedException $e) {
+            $this->assertEquals(
+                <<<EOF
+Failed asserting that object of class "stdClass" has attribute "privateAttribute".
+
+EOF
+                ,
+                TestFailure::exceptionToString($e)
+            );
+
+            return;
+        }
+
+        $this->fail();
+    }
+
+    public function testConstraintObjectHasAttribute2()
+    {
+        $constraint = new ObjectHasAttribute('privateAttribute');
+
+        try {
+            $constraint->evaluate(new \stdClass, 'custom message');
+        } catch (ExpectationFailedException $e) {
+            $this->assertEquals(
+                <<<EOF
+custom message
+Failed asserting that object of class "stdClass" has attribute "privateAttribute".
+
+EOF
+                ,
+                TestFailure::exceptionToString($e)
+            );
+
+            return;
+        }
+
+        $this->fail();
+    }
+}
